Lesson 3 (Week 2) Exercise

+++ www/js/controllers.js
@@ -222,19 +222,29 @@ angular.module('conFusion.controllers', [])
 
   }])
 
-  .controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate', function ($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate) {
+  .controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate', '$ionicPopup', '$ionicLoading', '$timeout', function ($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate, $ionicPopup, $ionicLoading, $timeout) {
 
     $scope.baseURL = baseURL;
     $scope.shouldShowDelete = false;
 
+    $ionicLoading.show({
+      template: '<ion-spinner></ion-spinner> Loading...'
+    })
+    
     $scope.favorites = favoriteFactory.getFavorites();
 
     $scope.dishes = menuFactory.getDishes().query(
       function (response) {
         $scope.dishes = response;
+        $timeout(function(){
+          $ionicLoading.hide()
+        }, 1000);
       },
       function (response) {
         $scope.message = "Error: " + response.status + " " + response.statusText;
+        $timeout(function(){
+          $ionicLoading.hide()
+        }, 1000);
       });
     console.log($scope.dishes, $scope.favorites);
 
@@ -244,21 +254,35 @@ angular.module('conFusion.controllers', [])
     }
 
     $scope.deleteFavorite = function (index) {
+      var confirmPopup = $ionicPopup.confirm({
+        title: 'Confirm Delete',
+        template: 'Are you sure you want to delete this item?'
+      });
 
-      favoriteFactory.deleteFromFavorites(index);
+      confirmPopup.then(function(res){
+        if(res){
+          console.log('Ok to delete');
+          favoriteFactory.deleteFromFavorites(index);
+        }
+        else {
+          console.log('Canceled delete');
+        }
+      })
+
+
       $scope.shouldShowDelete = false;
 
     }}])
 
   .filter('favoriteFilter', function () {
-    return function (dishes, favorites) {
-      var out = [];
-      for (var i = 0; i < favorites.length; i++) {
-        for (var j = 0; j < dishes.length; j++) {
-          if (dishes[j].id === favorites[i].id)
-            out.push(dishes[j]);
-        }
+  return function (dishes, favorites) {
+    var out = [];
+    for (var i = 0; i < favorites.length; i++) {
+      for (var j = 0; j < dishes.length; j++) {
+        if (dishes[j].id === favorites[i].id)
+          out.push(dishes[j]);
       }
-      return out;
+    }
+    return out;
   }});
 ;
